Move grpchelper into slam package
